Jonrad/config file (#2)
Added support for config files

+++ src/lsp.ts
@@ -3,56 +3,65 @@ import * as rpc from "vscode-jsonrpc";
 import { StreamMessageReader, StreamMessageWriter } from "vscode-jsonrpc/node";
 import { InitializeRequest } from "vscode-languageserver-protocol";
 import * as protocol from "vscode-languageserver-protocol";
-import { ClientCapabilities } from "vscode-languageserver-protocol/lib/common/protocol";
-import { consoleLogger } from "./logger";
 import { Logger } from "vscode-jsonrpc";
 import path from "path";
 
-const URI_SCHEME = "lsp";
-
 export interface LspClient {
+  id: string;
+  languages: string[];
+  extensions: string[];
+  start(): Promise<void>;
   dispose: () => void;
   sendRequest(method: string, args: any): Promise<any>;
   sendNotification(method: string, args: any): Promise<void>;
 }
 
-function buildUri(...paths: string[]) {
-  return `${URI_SCHEME}://` + path.resolve(...paths);
-}
+export class LspClientImpl implements LspClient {
 
-export async function startLsp(
-  command: string,
-  args: string[],
-  logger: Logger = consoleLogger,
-): Promise<LspClient> {
-  return LspClientImpl.create(command, args, logger);
-}
+  protected childProcess: ChildProcess | undefined;
 
-class LspClientImpl implements LspClient {
-  public static async create(
-    command: string,
-    args: string[],
-    logger: Logger,
-  ): Promise<LspClient> {
-    const childProcess = spawn(command, args);
-    const connection = rpc.createMessageConnection(
+  protected connection: rpc.MessageConnection | undefined;
+
+  protected capabilities: protocol.ServerCapabilities | undefined;
+
+  public constructor(
+    public readonly id: string,
+    public readonly languages: string[],
+    public readonly extensions: string[],
+    private readonly command: string,
+    private readonly args: string[],
+    private readonly logger: Logger, // TODO: better long term solution for logging
+  ) {}
+
+  public async start() {
+    if (this.isStarted()) {
+      return;
+    }
+
+    const childProcess = this.childProcess = spawn(this.command, this.args);
+
+    if (!childProcess.stdout || !childProcess.stdin) {
+      throw new Error("Child process not started");
+    }
+
+    const connection = this.connection = rpc.createMessageConnection(
       new StreamMessageReader(childProcess.stdout),
       new StreamMessageWriter(childProcess.stdin),
-      logger,
+      this.logger,
     );
 
     connection.onError((error) => {
-      logger.error(`Connection error: ${error}`);
+      this.logger.error(`Connection error: ${error}`);
       childProcess.kill();
     });
 
     connection.onClose(() => {
-      logger.log("Connection closed");
+      this.logger.log("Connection closed");
       childProcess.kill();
     });
 
     connection.onUnhandledNotification((notification) => {
-      logger.log(`Unhandled notification: ${JSON.stringify(notification)}`);
+      this.logger.log(`Unhandled notification: ${JSON.stringify(notification)}`);
     });
 
     connection.listen();
@@ -62,39 +71,47 @@ class LspClientImpl implements LspClient {
 
     const response = await connection.sendRequest(InitializeRequest.type, {
       processId: process.pid,
-      rootUri: buildUri('/'),
+      rootUri: "file:///", // TODO: figure out the best way to handle this
       capabilities: capabilities,
     });
 
-    logger.info(`Server LSP capabilities: ${JSON.stringify(response.capabilities, null, 2)}`);
-
-    return new LspClientImpl(
-      childProcess,
-      connection,
-      response.capabilities,
-      logger,
-    );
+    this.logger.info(`Server LSP capabilities: ${JSON.stringify(response.capabilities, null, 2)}`);
   }
 
-  private constructor(
-    private readonly childProcess: ChildProcess,
-    private readonly connection: rpc.MessageConnection,
-    private readonly capabilities: protocol.ServerCapabilities, // TODO: not sure what I'm doing with this, but it'll be needed I feel like
-    private readonly logger: Logger, // TODO: better long term solution for logging
-  ) {}
-
-  sendRequest(method: string, args: any): Promise<any> {
-    return this.connection.sendRequest(method, args);
+  private isStarted(): this is LspClientImpl & { connection: rpc.MessageConnection } {
+    return !!this.connection;
   }
 
-  sendNotification(method: string, args: any): Promise<void> {
-    return this.connection.sendNotification(method, args);
+  private assertStarted(): asserts this is LspClientImpl & { connection: rpc.MessageConnection } {
+    if (!this.connection) {
+      throw new Error("Not started");
+    }
+  }
+
+  async sendRequest(method: string, args: any): Promise<any> {
+    if (!this.isStarted()) {
+      await this.start();
+    }
+
+    this.assertStarted();
+
+    return await this.connection.sendRequest(method, args);
+  }
+
+  async sendNotification(method: string, args: any): Promise<void> {
+    if (!this.isStarted()) {
+      await this.start();
+    }
+
+    this.assertStarted();
+
+    return await this.connection.sendNotification(method, args);
   }
 
   dispose() {
     try {
-      this.connection.dispose();
-      this.childProcess.kill();
+      this.connection?.dispose();
+      this.childProcess?.kill();
     } catch (e: any) {
       this.logger.error(e.toString?.());
     }
